rag-service/requirements.txt
fastapi==0.110.2
uvicorn==0.29.0
requests==2.32.3
numpy==1.26.4
faiss-cpu==1.8.0
sentence-transformers==2.7.0
pydantic==2.6.4

minio
